+++ packages/sensors_plus/sensors_plus/android/src/main/kotlin/dev/fluttercommunity/plus/sensors/StreamHandlerImpl.kt
@@ -1,0 +1,42 @@
+package dev.fluttercommunity.plus.sensors
+
+import android.hardware.Sensor
+import android.hardware.SensorEvent
+import android.hardware.SensorEventListener
+import android.hardware.SensorManager
+import io.flutter.plugin.common.EventChannel
+import io.flutter.plugin.common.EventChannel.EventSink
+
+internal class StreamHandlerImpl(
+        private val sensorManager: SensorManager,
+        sensorType: Int
+) : EventChannel.StreamHandler {
+    private lateinit var sensorEventListener: SensorEventListener
+
+    private val sensor: Sensor by lazy {
+        sensorManager.getDefaultSensor(sensorType)
+    }
+
+    override fun onListen(arguments: Any?, events: EventSink) {
+        sensorEventListener = createSensorEventListener(events)
+        sensorManager.registerListener(sensorEventListener, sensor, SensorManager.SENSOR_DELAY_NORMAL)
+    }
+
+    override fun onCancel(arguments: Any?) {
+        sensorManager.unregisterListener(sensorEventListener)
+    }
+
+    private fun createSensorEventListener(events: EventSink): SensorEventListener {
+        return object : SensorEventListener {
+            override fun onAccuracyChanged(sensor: Sensor, accuracy: Int) {}
+
+            override fun onSensorChanged(event: SensorEvent) {
+                val sensorValues = DoubleArray(event.values.size)
+                event.values.forEachIndexed { index, value ->
+                    sensorValues[index] = value.toDouble()
+                }
+                events.success(sensorValues)
+            }
+        }
+    }
+}