+++ packages/package_info_plus/package_info_plus/CHANGELOG.md
@@ -1,3 +1,9 @@
+## 1.4.0
+
+- Android: Migrate to Kotlin
+- Android: Update dependencies, build config updates
+- Android: Fix project title
+
 ## 1.3.1
 
 - Fix embedding issue in example